Refactor EquationTool to single export
