+++ src/index.ts
@@ -22,67 +22,10 @@ async function main() {
     process.exit(1);
   }
 }
-import { ToolDefinition, ExecuteToolResponseSchema } from '@anthropic-ai/mcp-typescript-sdk';
-import { generateEquation } from './extensions/equation/equation-generator';
+import { equationTool as importedEquationTool } from './extensions/equation/equation-generator';
 
-// Define the tool
-const EquationTool: ToolDefinition = {
-  name: 'generate-latex-equation',
-  description: 'Converts natural language descriptions into properly formatted LaTeX equations',
-  parameters: {
-    type: 'object',
-    properties: {
-      description: {
-        type: 'string',
-        description: 'Natural language description of the equation to generate',
-      },
-      format: {
-        type: 'string',
-        enum: ['inline', 'display', 'align', 'gather', 'multline'],
-        description: 'LaTeX environment to use for the equation',
-        default: 'display',
-      },
-      numbered: {
-        type: 'boolean',
-        description: 'Whether the equation should be numbered',
-        default: false,
-      },
-      additionalContext: {
-        type: 'string',
-        description: 'Additional context to help with equation generation',
-      },
-    },
-    required: ['description'],
-  },
-  returns: {
-    type: 'object',
-    properties: {
-      latex: {
-        type: 'string',
-        description: 'Generated LaTeX code for the equation',
-      },
-      preview: {
-        type: 'string',
-        description: 'Plain text preview of how the equation might render',
-      },
-      explanation: {
-        type: 'string',
-        description: 'Explanation of the equation',
-      },
-    },
-  },
-};
-
-// Handler function
-async function handler(params: any): Promise<ExecuteToolResponseSchema> {
-  return generateEquation(params);
-}
-
-// Export the equation generation tool and handler
-export const equationTool = {
-  toolDefinition: EquationTool,
-  handler,
-};
+// Re-export the equation generation tool
+export const equationTool = importedEquationTool;
 // Only run the main function when this file is executed directly
 if (require.main === module) {
   main();
